% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idm_data_generation.R
\name{apply_iddata_censoring}
\alias{apply_iddata_censoring}
\title{Apply censoring times to an existing Illness-Death dataframe.}
\usage{
apply_iddata_censoring(iddata_df, censoring_times)
}
\arguments{
\item{iddata_df}{An Illness-Death dataframe in \code{iddata} format.}

\item{censoring_times}{Censoring time(s), either a single value or a vector
length \code{nrow(iddata_df)}}
}
\value{
An iddata tibble with updated censoring times
}
\description{
Update event times (Zt and Tt) and censoring indicators according to the
prescribed censoring time(s). Note that previous censoring information will
be updated where newer censoring times are provided.
}
