% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idm_data_generation.R
\name{apply_iddata_death}
\alias{apply_iddata_death}
\title{Update death times in an existing Illness-Death dataframe.}
\usage{
apply_iddata_death(iddata_df, death_times)
}
\arguments{
\item{iddata_df}{An Illness-Death dataframe in \code{iddata} format.}

\item{death_times}{Updated death time(s), either a single value or a vector
length \code{nrow(iddata_df)}}
}
\value{
An iddata tibble with updated death times
}
\description{
Update event times (Zt and Tt) according to the prescribed death time(s).
Death times are only updated when the new death time is earlier than the
previous one for a given individual. This function cannot postpone an
existing death time as the parameters provide no way to know whether illness
occurred between old and new death time.
}
