% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survexp_wrappers.R
\name{.filter_times}
\alias{.filter_times}
\title{Map time steps to event times}
\usage{
.filter_times(time_steps, event_times)
}
\arguments{
\item{time_steps}{An array/vector of desired timesteps}

\item{event_times}{An array/vector of actual event-times or breakpoints}
}
\value{
An array of unique time points ordered in increasing order
}
\description{
Filter and map a list/vector of time steps to a list/vector of event times.
This is useful for evaluating a piecewise constant function (e.g Kaplan-Meier survival, binomial regression)
but limit the number of redundant computation steps, by evaluating the function only at breakpoints.
}
\examples{
.filter_times(c(1,2,4,5,6),c(1,3,5))
}
