% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idm_data_generation.R
\name{generate_exponential_time_to_event}
\alias{generate_exponential_time_to_event}
\title{Generate time to events from an exponential distribution}
\usage{
generate_exponential_time_to_event(n_individuals, lambda)
}
\arguments{
\item{n_individuals}{integer, Number of individuals to simulate}

\item{lambda}{float > 0, Rate of the underlying Poisson point process}
}
\value{
Atomic vector of time to (first) event.
}
\description{
This is a simple wrapper around the stats::rexp function, without added value
beyond consistent argument checking and format with other generating functions
from the package.
}
