% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idm_data_generation.R
\name{generate_poisson_time_to_event}
\alias{generate_poisson_time_to_event}
\title{Generate synthetic time to event data from a homogeneous Poisson process.}
\usage{
generate_poisson_time_to_event(
  n_individuals,
  lambda,
  n_timesteps,
  recurring = FALSE
)
}
\arguments{
\item{n_individuals}{integer, Number of individuals to simulate}

\item{lambda}{float > 0, Rate of the Poisson point process}

\item{n_timesteps}{integer, Number of evenly spaced timesteps}

\item{recurring}{logical, whether event is recurring or fully absorbing.}
}
\value{
If recurring=FALSE a vector of first time to event, if recurring=TRUE
a list of event times of length n_individuals.
}
\description{
Note that this is not a efficient way to model time homogeneous non recurring
events that may be better captured by a continuous time exponential
distribution from \code{\link{generate_exponential_time_to_event}}.
}
