% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idm_data_generation.R
\name{generate_uncensored_ind_exp_idm_data}
\alias{generate_uncensored_ind_exp_idm_data}
\title{Generate uncensored independent exponential illness-death data.}
\usage{
generate_uncensored_ind_exp_idm_data(
  n_individuals,
  lambda_illness,
  lambda_death
)
}
\arguments{
\item{n_individuals}{integer, Number of individuals to simulate}

\item{lambda_illness}{float > 0, Rate of the underlying Poisson point process
to transition from healthy to illness}

\item{lambda_death}{float > 0, Rate of death, either from healthy or illness
state.}
}
\value{
a tibble of iddata
}
\description{
Mostly for testing purposes. Illness and death times are assumed independent,
in particular the death rate is assumed equal in healthy and illness states.
This simplification creates data that are not really following an
Illness-Death model but a simpler semi competitive model, where death
prevents observation of illness times.
}
