% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewnetTPreg.R
\name{get_survival_at}
\alias{get_survival_at}
\title{Estimated survival at time t from a survfit like df.}
\usage{
get_survival_at(t, survfit_data, safe = TRUE)
}
\arguments{
\item{t}{Single or vector numeric value $\geq$ 0.}

\item{survfit_data}{A survfit, summary.survfit or data.frame object. If
data.frame, it should contain at least columns \code{surv} and \code{time}.}

\item{safe}{boolean, Indicate whether to enable input argument checking,
default to TRUE.}
}
\value{
A survival probabilities of same length as \code{t}.
}
\description{
Estimated survival at time t from a survfit like df.
}
