\name{idmTPreg-package}
\alias{idmTPreg-package}
\alias{idmTPreg}
\docType{package}
\title{
\packageTitle{idmTPreg}
}
\description{
\packageDescription{idmTPreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{idmTPreg}
\packageIndices{idmTPreg}
}
\author{
\packageAuthor{idmTPreg}

Maintainer: \packageMaintainer{idmTPreg}
}
\references{
Azarang, L. Scheike, TH. and de Una-Alvarez, J. (2017) \emph{Direct modeling of regression effects for transition probabilities
in the progressive illness-death model} Statistics in Medicine \bold{36}, \eqn{1964-1976}. 
}

