% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survexp_wrappers.R
\name{indiv_survprob_pch}
\alias{indiv_survprob_pch}
\title{Compute expected survival for one individual at several times}
\usage{
indiv_survprob_pch(individual_df, eval_times, ratetable, rmap, fast = FALSE)
}
\arguments{
\item{individual_df}{A DF or DT containing a single row with a single patient's covariates}

\item{eval_times}{Times at which the expected survival should be evaluated. Times must be in days if using a ratetable with a Date component.}

\item{ratetable}{A ratetable object}

\item{rmap}{An rmap argument as the one passed to survival::survexp. This argument will only be evaluated inside survexp}

\item{fast}{Binary, if TRUE the function will omit some validity checks on the arguments. Should be used with caution.}
}
\value{
Returns a tibble with two columns: eval_times and surv the survival probability from the survexp object at those times. Upon failure of surv.exp because of missing data in the DF row, a vector of NA will be returned.
}
\description{
A wrapper function calling survival::survexp to compute expected survival probablity of a single individual for one or several times, using a piecewise constant hazard (pch) defined in a ratetable object.
}
