% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{invalid_argument}
\alias{invalid_argument}
\title{Custom exception for invalid arguments}
\usage{
invalid_argument(arg, must, not = NULL)
}
\arguments{
\item{arg}{string, faulty argument name}

\item{must}{string, description of the expected nature of the argument}

\item{not}{string, optional, description of the received argument}
}
\value{
None
}
\description{
Derived from https://adv-r.hadley.nz/conditions.html#custom-conditions
}
