% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewnetTPreg.R
\name{mod.glm.fit.callingwrapper}
\alias{mod.glm.fit.callingwrapper}
\title{Check input and output of mod.glm.fit}
\usage{
mod.glm.fit.callingwrapper(
  X,
  response,
  family,
  weights,
  maxit = stats::glm.control()$maxit,
  maxmaxit = 1000,
  warning_str = "",
  ...
)
}
\arguments{
\item{X}{array, must have at least 2 dimensions}

\item{...}{}
}
\value{
A set of glm coefficients
}
\description{
Avoids calling the glm fitting function when there are no contrasted response
values (all 0 or all 1). Ensures return of NA coefficient when the algorithm
did not converge.
}
