% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewnetTPreg.R
\name{mod.glm.fit.errorwrapper}
\alias{mod.glm.fit.errorwrapper}
\title{Wraps the call to mod.glm.fit to handle convergence errors}
\usage{
mod.glm.fit.errorwrapper(
  X,
  response,
  family,
  weights,
  maxit = stats::glm.control()$maxit,
  maxmaxit = 1000,
  warning_str = "",
  ...
)
}
\arguments{
\item{...}{}
}
\value{
A glm object
}
\description{
Tries to handle convergence and other issues upon calling the mod.glm.fit
function. Some errors returned by mod.glm.fit are cast to warnings. In order
to reduce crashes several values of \code{maxit} are tested if convergence issues
are detected. NAs will be returned in case of crash/divergence or lack of
informative observations.
}
