% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netidmtpreg-package.R
\docType{package}
\name{netidmtpreg-package}
\alias{netidmtpreg}
\alias{netidmtpreg-package}
\title{netidmtpreg: Net Survival Binomial Regression for Progressive Illness Death Models}
\description{
This package implements utilities to infer and inspect transition probabilities in a progressive illness-death model in a net survival context through binomial regression. As such it also provides routines to analyze illness-death transition probabilities without correcting for background mortality. Ref: Azarang L & Giorgi R. (2021) \doi{10.1177/09622802211003608}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qmarcou/netidmtpreg}
  \item Report bugs at \url{https://github.com/qmarcou/netidmtpreg/issues}
}

}
\author{
\strong{Maintainer}: Quentin Marcou \email{quentin.marcou@univ-amu.fr} (\href{https://orcid.org/0000-0001-7074-2761}{ORCID})

}
\keyword{internal}
