% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewnetTPreg.R
\name{offsetlogit}
\alias{offsetlogit}
\title{An offset survival logit link function closure for 13 and 23 transitions.}
\usage{
offsetlogit(s, t, data_df, ratetable, rmap)
}
\arguments{
\item{s}{positive scalar.}

\item{t}{positive scalar.}

\item{data_df}{A \code{data.frame} containing population characteristics.}

\item{ratetable}{a \code{ratetable} object.}

\item{rmap}{see survival::survexp.fit.}
}
\value{
a \code{link-glm} object
}
\description{
Returns a offset logit \code{link-glm} object taking into account population
mortality a the considered times s and t.
}
\examples{
# ADD_EXAMPLES_HERE
}
