% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewnetTPreg.R
\name{renewnetTPreg}
\alias{renewnetTPreg}
\title{Fit (net) survival Illness-Death transition probabilities}
\usage{
renewnetTPreg(
  s = 0,
  t = NULL,
  trans,
  formula,
  ratetable,
  time_dep_popvars = NULL,
  rmap = NULL,
  data,
  link = "logit",
  R = 199,
  by = NULL,
  readjust_t = TRUE
)
}
\arguments{
\item{s}{DESCRIPTION.}

\item{t}{Times at which survival should be estimated, either \code{NULL}, scalar
or vector of numerics. Default to NULL. If a scalar value is given, it will
be interpreted as the maximum time at which survival shall be estimated
using a non-parametric approach. If a vector or list-like object, the values
will be interpreted as the list of times at which one whishes to get point
estimations of survivals. By default estimation will be made at the time of
last events before the requested times. This behavior can be altered by
setting \code{readjust_t} to \code{FALSE}, mostly for testing purposes. If set to
\code{NULL} (default) \code{t} is set as a scalar value to the last observed
transition time.}

\item{trans}{DESCRIPTION.}

\item{formula}{DESCRIPTION.}

\item{ratetable}{DESCRIPTION.}

\item{time_dep_popvars}{DESCRIPTION.}

\item{rmap}{DESCRIPTION.}

\item{data}{DESCRIPTION.}

\item{link}{DESCRIPTION.}

\item{R}{DESCRIPTION.}

\item{by}{DESCRIPTION.}

\item{readjust_t}{boolean, default \code{TRUE}. Whether to use exact \code{t} values
(\code{FALSE}) or adjust t values to observed transition times (\code{TRUE}).}
}
\value{
RETURN_DESCRIPTION
}
\description{
Core function of the package, implementing a
}
\examples{
# ADD_EXAMPLES_HERE
}
\references{
Azarang, L., Scheike, T., & de Uña‐Álvarez, J. (2017). Direct modeling of
regression effects for transition probabilities in the progressive
illness–death model. \strong{Statistics in Medicine}, 36(12), 1964-1976.

Azarang L, Giorgi R. (2021). Estimation of covariate effects on net
survivals in the relative survival progressive illness-death model.
\strong{Statistical Methods in Medical Research}, 30(6), 1538-1553.
doi:10.1177/09622802211003608
}
